Исправить ошибки валидации

+++ main-service/src/main/java/ru/practicum/ewm/admin/service/AdminCategoryServiceImpl.java
@@ -4,15 +4,18 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.dto.CategoryDto;
+import ru.practicum.ewm.exception.DataConflictException;
 import ru.practicum.ewm.exception.NotFoundException;
 import ru.practicum.ewm.mapper.CategoryMapper;
 import ru.practicum.ewm.model.Category;
 import ru.practicum.ewm.repository.CategoryRepository;
+import ru.practicum.ewm.repository.EventRepository;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class AdminCategoryServiceImpl implements AdminCategoryService {
+    private final EventRepository eventRepository;
     private final CategoryRepository categoryRepository;
 
     @Override
@@ -25,6 +28,10 @@ public class AdminCategoryServiceImpl implements AdminCategoryService {
     @Override
     public void deleteCategory(Long catId) {
         validateCategory(catId);
+        if (eventRepository.countByCategoryId(catId) > 0) {
+            throw new DataConflictException("Некорректные условия операции",
+                    "У категории с id = " + catId + " есть связанные события");
+        }
         categoryRepository.deleteById(catId);
     }
